refactor: product card를 별도 ui로 분리
